+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 0.30.2
+## 0.30.3
 * small updates to `Settings`
 * fixed a bug in path to json files in `FileDataManager`
+* improvement to saving task list across app re-start
 
 ## 0.30.1
 * `DebugLevel` is now an enum